--- source/robotis_lab/robotis_lab/tasks/manager_based/OMY/stack/ik_rel_env_cfg.py
@@ -1,52 +1,0 @@
-# Copyright 2025 ROBOTIS CO., LTD.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-# Author: Taehyeong Kim
-
-# Copyright (c) 2022-2025, The Isaac Lab Project Developers.
-# All rights reserved.
-#
-# SPDX-License-Identifier: BSD-3-Clause
-
-from isaaclab.controllers.differential_ik_cfg import DifferentialIKControllerCfg
-from isaaclab.envs.mdp.actions.actions_cfg import DifferentialInverseKinematicsActionCfg
-from isaaclab.utils import configclass
-
-from . import joint_pos_env_cfg
-
-##
-# Pre-defined configs
-##
-from robotis_lab.assets.OMY import OMY_CFG  # isort: skip
-
-
-@configclass
-class OMYCubeStackEnvCfg(joint_pos_env_cfg.OMYCubeStackEnvCfg):
-    def __post_init__(self):
-        # post init of parent
-        super().__post_init__()
-
-        # Set OMY as robot
-        # We switch here to a stiffer PD controller for IK tracking to be better.
-        self.scene.robot = OMY_CFG.replace(prim_path="{ENV_REGEX_NS}/Robot")
-
-        # Set actions for the specific robot type (OMY)
-        self.actions.arm_action = DifferentialInverseKinematicsActionCfg(
-            asset_name="robot",
-            joint_names=["joint[1-6]"],
-            body_name="link6",
-            controller=DifferentialIKControllerCfg(command_type="pose", use_relative_mode=True, ik_method="dls"),
-            scale=0.1,
-            body_offset=DifferentialInverseKinematicsActionCfg.OffsetCfg(pos=[0.0, 0.0, 0.0]),
-        )
